Change ViewModel from ObservableObject to Observable
